use atomic boolean + instrument service method
