+++ apm-agent-plugins/apm-servlet-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.bci.ElasticApmInstrumentation
@@ -1,5 +1,6 @@
 co.elastic.apm.agent.servlet.ServletInstrumentation
-co.elastic.apm.agent.servlet.ServletInitInstrumentation
+co.elastic.apm.agent.servlet.ServletVersionInstrumentation$Init
+co.elastic.apm.agent.servlet.ServletVersionInstrumentation$Service
 co.elastic.apm.agent.servlet.FilterChainInstrumentation
 co.elastic.apm.agent.servlet.AsyncInstrumentation$StartAsyncInstrumentation
 co.elastic.apm.agent.servlet.AsyncInstrumentation$AsyncContextInstrumentation
